requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
httpx>=0.27.0
python-dotenv>=1.0.1
apscheduler>=3.10.4
pydantic>=2.0.0